Generate Achievements

External function to generate achievements for HEMA ratings. This function retrieves data from the specified path or uses package data as a fallback. It computes achievements by invoking all functions prefixed with `ach_` in the namespace of the `HEMAscRaper` package.

generate_achievements(path = "./hema_ratings", export_csv = FALSE)

Arguments

path

Character. Path to the folder containing HEMA data files. Defaults to "./hema_ratings".

export_csv

Logical. If TRUE, exports the achievements list as a CSV to path.

Value

A data frame containing all achievements, with columns:

fighter_id

ID of the fighter.

tier_id

Numeric tier ID (1-4).

achieved

Logical, TRUE if achieved.

percentile

Proportion of fighters who achieved this tier.

achievement_tier

Character. Tier name ("Bronze", "Silver", etc.).

achievement_name

Character. Name of the achievement.

achievement_description

Character. Description of the achievement.

achievement_icon

Character. Icon file name for the achievement.

Examples

if (FALSE) { # \dontrun{
generate_achievements("./hema_ratings")
} # }