Download and Extract HEMA Fights Data by Year

This function downloads data from the *HEMA Ratings* website and extracts detailed fight statistics for events that occurred within a specified year or set of years. If no year is specified, the function gathers fight data for all available events.

get_fights(year = NULL, raw = FALSE)

Arguments

year

An optional vector of years (as integers) to filter the events. If `NULL`, data for all available years is fetched. Defaults to `NULL`.

raw

Logical. If `TRUE`, returns the raw, unprocessed data as scraped from the website. If `FALSE` (default), the data is normalized for consistency (e.g., names, categories, and stages).

Value

A tibble containing detailed fight statistics for the specified events with columns that may include:

event_id

Unique identifier of the event. Integer.

event_name

Name of the event. Character.

tournament_name

Name of the specific tournament or discipline within the event. Character.

tournament_category

Category of the tournament (e.g., Mixed, Men's, Women's, Underrepresented Genders). Character.

event_weapon

Weapon used in the tournament (e.g., Longsword, Rapier). Character.

fighter_id

Unique identifier of the first fighter in a match. Integer.

opponent_id

Unique identifier of the opponent fighter in a match. Integer.

fighter_score

Score of the first fighter. Integer.

opponent_score

Score of the opponent fighter. Integer.

Details

This function scrapes the HEMA Ratings events page and compiles a list of all fights for the specified years. If no `year` is provided, it extracts fight data for all events available on the website. For each event, it calls the `.get_tournament()` function to gather detailed fight-level information.

When the `raw` parameter is set to `FALSE`, the function normalizes data for consistency, ensuring uniformity in names, categories, and other attributes. Missing data (e.g., anonymized fighters) is handled appropriately.

Note

The function uses `rvest` for HTML parsing and `purrr` for handling multiple event pages. It performs basic error handling to check if events matching the specified years are available. Depending on the number of events and the size of the tournament data, the function may take some time to run.

Examples

if (FALSE) { # \dontrun{
# Download all HEMA fight data for the year 2022
fights_2022 <- get_fights(2022)

# Download HEMA fight data for multiple years (e.g., 2019 and 2020)
fights_2019_2020 <- get_fights(c(2019, 2020))

# Download HEMA fight data for all available years
all_fights <- get_fights()

# Download raw fight data for 2022 without normalization
raw_fights_2022 <- get_fights(2022, raw = TRUE)
} # }