Visualize HEMA Communities

Generates a visual representation of a graph object, applying community detection and customizable layouts. Filters edges by weight and applies a color palette to distinguish communities. Provides an aesthetically styled plot.

visualize_communities(
  graph,
  color_palette = NULL,
  layout_type = "fr",
  weight_threshold_dotted = 5,
  show_labels = TRUE
)

Arguments

graph

A `tbl_graph` object from the `tidygraph` package, representing nodes and edges of the graph.

color_palette

A character vector specifying custom colors for communities. If `NULL`, a default palette is used.

layout_type

A character string specifying the layout algorithm. Options include `"fr"` (default), `"kk"`, `"mds"`, and `"circle"`.

weight_threshold_dotted

A numeric value specifying the threshold for dotted line edges. Default: 5.

show_labels

boolean - turn off the grapgh description.

Value

A ggplot object representing the visualized graph.

Examples

if (FALSE) { # \dontrun{
# Example 1: Basic graph visualization with default settings
library(dplyr)
generate_communities_graph( tournament_weapon = "Steel Longsword", 
                            year=2024,
                           weight_threshold = 5, 
                           population_threshold = 15)%>%
visualize_communities()

# Example 2: Custom layout and color palette
custom_palette <- c("#1B9E77", "#D95F02", "#7570B3")
visualize_communities(graph, year = 2024, color_palette = custom_palette, layout_type = "kk")
} # }